requirements.txt
boto3==1.26.69
botocore==1.29.69
Pillow==9.3